refactored pseudomaterial generation methods
